requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
pywebview==4.4.1
werkzeug==2.3.7
sqlite3 